Created subproject for lidar perih
Idea is to have 1 mcu and 9 vl53l0x lidars oriented in different directions.
The mcu will configure all sensors and collect the data before sending it to AP.

The reason for not connecting them directly to AP is because  when using multiple vl53* sensors
their I2C address must be changed.
For this to work, the XSHUT pin must be pulled low on all but the current one.
While this would be possible from Ap with lua scripting, it would require alot of "wasted" GPIOs, so
it seemed more logical to just add an mcu that handles it all.